Fix & Feat: method update

+++ internal/dto/directory_tree/create.go
@@ -1,13 +1,13 @@
 package dto
 
 type CreateDirectoryTreeReq struct {
-	ParentID int64  `json:"parentId"`
+	ParentID int32  `json:"parentId"`
 	Name     string `json:"name" validate:"required"`
 }
 
 type CreateDirectoryTreeRes struct {
-	Id          int64  `json:"id"`
-	DirectoryId int64  `json:"directoryId"`
-	ParentID    int64  `json:"parentId"`
+	Id          int32  `json:"id"`
+	DirectoryId int32  `json:"directoryId"`
+	ParentID    int32  `json:"parentId"`
 	Name        string `json:"name`
 }
